refactor(frontend): move primary-docs form to seperate component
